# docs/aquatrack_1_6_openapi.yaml
openapi: 3.0.3
info:
  title: AquaTrack 1.6 MVP API
  version: "1.6.0"
  description: >
    AquaTrack / 游跡 — 最小可行產品（MVP）API。
    主要涵蓋：Auth、User、CompetitionResults、PB、Teams、Join Codes、Plans、Targets、Leaderboards。
servers:
  - url: https://api.aquatrack.example.com/v1
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Email 登入
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登入成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /me:
    get:
      tags: [Users]
      summary: 取得我的基本資料
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      tags: [Users]
      summary: 更新我的基本資料
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Unauthorized:
      description: 未授權
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
...
    AuthTokens:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: Bearer }
        expires_in: { type: integer, example: 3600 }
    User:
      type: object
      properties:
        user_id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [swimmer, coach, admin] }
        created_at: { type: string, format: date-time }
    UserUpdateRequest:
      type: object
      properties:
        name: { type: string }
...